Release Prep 2.0.0 (#201)
* Remove Pods / Code Readability Refactor (#199)

* refactor: removing PodSpec file as RNOrientation no longer support pods

* refactor: readability of package entry point file

* chore: adding contributors, cleaning up package.json and updating preffered depdendencies to RN 40+

* release: preparing for 2.0.0 release

* Update Demo Example App (#193)

* Updated RN to 0.45

* Updated demo

* Removed bind(this) in favor of class property shorthand

* Fixed white space and typo

* Refactor RN 45.1 Demo Update (#200)

* Updated RN to 0.45

* Updated demo

* Removed bind(this) in favor of class property shorthand

* Fixed white space and typo

* refactor: removing double code
