+++ demo/.babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": ["react-native"]
+}
