+++ demo/android/app/src/main/java/com/demo/MainActivity.java
@@ -1,47 +1,11 @@
 package com.demo;
 
+import com.facebook.react.ReactActivity;
+
 import android.content.Intent;
 import android.content.res.Configuration;
 
-import com.facebook.react.ReactActivity;
-import com.github.yamill.orientation.OrientationPackage;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-
-import java.util.Arrays;
-import java.util.List;
-
 public class MainActivity extends ReactActivity {
-
-    /**
-     * Returns the name of the main component registered from JavaScript.
-     * This is used to schedule rendering of the component.
-     */
-    @Override
-    protected String getMainComponentName() {
-        return "demo";
-    }
-
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    @Override
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-    /**
-     * A list of packages used by the app. If the app uses additional views
-     * or modules besides the default ones, add more packages here.
-     */
-    @Override
-    protected List<ReactPackage> getPackages() {
-        return Arrays.<ReactPackage>asList(
-            new MainReactPackage(),
-            new OrientationPackage(this)
-        );
-    }
 
     @Override
     public void onConfigurationChanged(Configuration newConfig) {
@@ -49,5 +13,13 @@ public class MainActivity extends ReactActivity {
         Intent intent = new Intent("onConfigurationChanged");
         intent.putExtra("newConfig", newConfig);
         this.sendBroadcast(intent);
+    }
+
+    /**
+     * Returns the name of the main component registered from JavaScript.
+     * This is used to schedule rendering of the component.
+     */
+    protected String getMainComponentName() {
+        return "demo";
     }
 }
