+++ demo/android/settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'demo'
-
-include ':app'
 include ':react-native-orientation'
 project(':react-native-orientation').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-orientation/android')
+
+include ':app'
