+++ demo/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "demo",
+  "displayName": "demo"
+}
