+++ demo/package.json
@@ -1,13 +1,16 @@
 {
-  "name": "demo",
-  "version": "0.0.1",
+  "name": "ReactNativeOrientationDemo",
+  "version": "2.0.0",
   "private": true,
   "scripts": {
-    "start": "node_modules/react-native/packager/packager.sh"
+    "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "^15.1.0",
-    "react-native": "^0.27.2",
-    "react-native-orientation": ".."
+    "react": "16.0.0-alpha.12",
+    "react-native": "^0.45.1",
+    "react-native-orientation": "file:.."
+  },
+  "devDependencies": {
+    "babel-preset-react-native": "^2.0.0"
   }
 }
